fix: move page state up component hierachy
Tracking needs to extract the page prop from author profile. However
the way that this was written meant that the author profile had a page
prop then further down the component hierachy the page state was managed
by the pagination helper which managed the page props of it's children.
This change moves the pagination helper up so that it manages page state
for the entire tree so author profile has the correct page number prop
to be used for tracking events.