fix: ensure consistent scrolling behaviour on author profile
The Next and Prev buttons should scroll to the top of the page when
clicked. Previously, they would scroll upwards, but either scroll too
far or not far enough. This was because the layout was recalculated
while the scroll was happening. This change delays the scroll until
after the layout so that it always scrolls to the top.